tests/wpt/web-platform-tests/mediacapture-image/getPhotoSettings.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/mediacapture-image/resources/imagecapture-helpers.js"></script>
<body>
<canvas id='canvas' width=10 height=10/>
</body>
<script>

const fillLightModeNames = ['off', 'auto', 'flash'];

// This test verifies that ImageCapture can call getPhotoSettings(), with a
// mock Mojo interface implementation.
image_capture_test(async (t, imageCaptureTest) => {
  let canvas = document.getElementById('canvas');
  let context = canvas.getContext('2d');
  context.fillStyle = 'red';
  context.fillRect(0, 0, 10, 10);
  let stream = canvas.captureStream();

  let mockState = imageCaptureTest.mockImageCapture().state();
  let capturer = new ImageCapture(stream.getVideoTracks()[0]);
  let settings = await capturer.getPhotoSettings();
  assert_equals(settings.imageWidth, mockState.width.current, 'width');
  assert_equals(settings.imageHeight, mockState.height.current, 'height');
  // TODO(mcasas): check the remaining two entries https://crbug.com/732521.

}, 'exercises ImageCapture.getPhotoSettings()');

</script>
